+++ arc-core/src/main/java/fr/insee/arc/core/service/global/thread/MultiThreading.java
@@ -15,12 +15,10 @@ import fr.insee.arc.core.service.global.scalability.ScalableConnection;
 import fr.insee.arc.core.service.mutiphase.ApiMultiphaseService;
 import fr.insee.arc.core.service.mutiphase.thread.ThreadMultiphaseService;
 import fr.insee.arc.core.util.StaticLoggerDispatcher;
-import fr.insee.arc.utils.dao.GenericPreparedStatementBuilder;
 import fr.insee.arc.utils.dao.UtilitaireDao;
 import fr.insee.arc.utils.database.ArcDatabase;
 import fr.insee.arc.utils.exception.ArcException;
 import fr.insee.arc.utils.exception.ArcExceptionMessage;
-import fr.insee.arc.utils.utils.FormatSQL;
 
 /**
  * class meant to start multithreading