+++ arc-core/src/main/java/fr/insee/arc/core/service/p6export/operation/ExportOperation.java
@@ -107,5 +107,11 @@ public class ExportOperation {
 	public void rollBack() throws ArcException {
 		exportParquetDao.rollback();
 	}
+
+	public String getParamBatch() {
+		return paramBatch;
+	}
+	
+	
 	
 }